Dev branch: ESPHome compatibility fixes
- Restructured component for proper ESPHome platform architecture
- Fixed CONFIG_SCHEMA deprecation warnings
- Updated trigger definitions for ESPHome 2025.7.4
- Created separate cover.py platform file
- Configuration validation now passes
- Ready for C++ compilation debugging

Status:
✅ Configuration validation passes
✅ All required files in correct structure
⚠️  C++ compilation needs debugging
- TypeError in iter_component_configs needs investigation
