+++ test-config.yaml
@@ -1,0 +1,81 @@
+esphome:
+  name: test-impulse-cover
+  friendly_name: Test Impulse Cover
+
+esp32:
+  board: esp32dev
+  framework:
+    type: arduino
+
+# Enable logging
+logger:
+  level: DEBUG
+
+# Enable Home Assistant API
+api:
+
+ota:
+
+wifi:
+  ssid: "test-network"
+  password: "test1234"
+  
+  # Enable fallback hotspot
+  ap:
+    ssid: "Test Fallback"
+    password: "test1234"
+
+captive_portal:
+
+# Import the custom component (local path for testing)
+external_components:
+  - source: 
+      type: local
+      path: components
+    components: [ impulse_cover ]
+
+# GPIO Output for sending pulses
+output:
+  - platform: gpio
+    pin: GPIO2
+    id: test_gate_output
+
+# Optional: Binary sensors for testing
+binary_sensor:
+  - platform: gpio
+    pin: 
+      number: GPIO4
+      mode: INPUT_PULLUP
+      inverted: true
+    name: "Test Gate Open"
+    id: test_gate_open
+    
+  - platform: gpio
+    pin:
+      number: GPIO5  
+      mode: INPUT_PULLUP
+      inverted: true
+    name: "Test Gate Close"
+    id: test_gate_close
+
+# Main cover configuration for testing
+cover:
+  - platform: impulse_cover
+    name: "Test Gate"
+    id: test_gate
+    
+    # Required: Output for sending pulses
+    output: test_gate_output
+    
+    # Required: Timing configuration
+    open_duration: 15s
+    close_duration: 15s
+    
+    # Optional: Safety and pulse timing
+    pulse_delay: 500ms
+    safety_timeout: 60s
+    safety_max_cycles: 5
+    
+    # Optional: Endstop sensors
+    open_sensor: test_gate_open
+    close_sensor: test_gate_close
